change relogin as guest when attempts the connexion with api42 + passport-42 become passport-oauth2 + protect userservice func deleteUser() + resize uploaded image comments + localStorage become sessionStorage + logout function from authContext.tsx to Account.tsx + redirection after guest logout

+++ backend/src/users/users.service.ts
@@ -66,7 +66,11 @@ export class UsersService {
   }
 
   async deleteUser(id: number): Promise<UsersEntity> {
+    if (!id)
+      return null;
     const user = await this.getById(id);
+    if (!user)
+      return null;
     return await this.userRepo.remove(user);
   }
